NEX-44: Move contact for to its own component
